requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
google-genai
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
markdown-pdf